feat: abstraindo as chamadas das actions
