+++ src/contexts/CyclesContext.tsx
@@ -1,5 +1,10 @@
 import { createContext, ReactNode, useReducer, useState } from "react";
-import { ActionTypes, Cycle, cyclesReducer } from "../reducers/cycles";
+import {
+	addNewCycleAction,
+	interruptCurrentCycleAction,
+	markCurrentCycleAsFinishedAction,
+} from "../reducers/cycles/actions";
+import { Cycle, cyclesReducer } from "../reducers/cycles/reducer";
 
 interface CreateCycleData {
 	task: string;
@@ -45,12 +50,7 @@ export function CyclesContextProvider({
 	}
 
 	function markCurrentCycleAsFinished() {
-		dispatch({
-			type: ActionTypes.MARK_CURRENT_CYCLE_AS_FINISHED,
-			payload: {
-				activeCycleId,
-			},
-		});
+		dispatch(markCurrentCycleAsFinishedAction());
 	}
 
 	/**
@@ -66,12 +66,7 @@ export function CyclesContextProvider({
 			startDate: new Date(),
 		};
 
-		dispatch({
-			type: ActionTypes.ADD_NEW_CYCLE,
-			payload: {
-				newCycle,
-			},
-		});
+		dispatch(addNewCycleAction(newCycle));
 
 		setAmountSecondsPassed(0);
 	}
@@ -88,12 +83,7 @@ export function CyclesContextProvider({
 	 * pra daí fazer a alteração.
 	 */
 	function interruptCurrentCycle() {
-		dispatch({
-			type: ActionTypes.INTERRUPT_CURRENT_CYCLE,
-			payload: {
-				activeCycleId,
-			},
-		});
+		dispatch(interruptCurrentCycleAction());
 	}
 
 	return (
