+++ src/reducers/cycles/reducer.ts
@@ -1,3 +1,5 @@
+import { ActionTypes } from "./actions";
+
 export interface Cycle {
 	id: string;
 	task: string;
@@ -10,12 +12,6 @@ export interface Cycle {
 interface CyclesState {
 	cycles: Cycle[];
 	activeCycleId: string | null;
-}
-
-export enum ActionTypes {
-	ADD_NEW_CYCLE = "ADD_NEW_CYCLE",
-	INTERRUPT_CURRENT_CYCLE = "INTERRUPT_CURRENT_CYCLE",
-	MARK_CURRENT_CYCLE_AS_FINISHED = "MARK_CURRENT_CYCLE_AS_FINISHED",
 }
 
 /**
@@ -31,7 +27,7 @@ export enum ActionTypes {
  * do código por outros desenvolvedores familiarizados com essa prática.
  */
 export function cyclesReducer(state: CyclesState, action: any) {
-	switch (action) {
+	switch (action.type) {
 		/**
 		 * Retorna um objeto contendo os estados anteriores, com um array de ciclos
 		 * e adiciona um novo ciclo ao final. Por fim, define o novo ciclo como ativo.
